+++ RELEASE_NOTES.md
@@ -1,5 +1,6 @@
 #### 1.0.0-beta006 - 04.03.2016
 * Fix XRoadRequest class to make POST requests instead of GET.
+* Refactor X-Road meta services to handle them separately.
 
 #### 1.0.0-beta005 - 04.03.2016
 * Fix Xop binary content serialization bug.